Migrating to Typescript following https://reactnative.dev/docs/typescript#adding-typescript-to-an-existing-project and following https://github.com/DefinitelyTyped/DefinitelyTyped/issues/33311#issuecomment-466731156 to fix issues reported when running 'yarn tsc' command
